pyproject.toml
[tool.poetry]
name = "avaaz-api"
version = "0.1.0"
description = "An Example of api"
authors = ["johnatasr <johnatsrs60@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "2.1.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.7.0"
Flask-Cors = "^3.0.10"
Flask-WTF = "^0.14.3"
SQLAlchemy = "^1.4.5"
pytest-mock = "^3.5.1"
flask-restx = "^1.0.5"
pymysql = "^1.0.2"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.1"
jinja2 = "^3.1.2"
python-dateutil = "^2.8.2"
python-dotenv = "^0.21.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.2"
coverage = "^5.5"
pytest = "^6.2.2"
autopep8 = "^1.5.6"
blinker = "^1.4"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
